#49: WIP: "dm volume delete" command and some plumbing.
It does a "DatameshRPC.DeleteVolume" call to the server.

That puts a deletion marker in etcd at
/datamesh.io/filesystems/deleted/FSID, and removes it from the registry.

This is noticed by all the nodes and injects a "delete" event to the
fsmachine.

That tries to "zfs destroy -r" the top-level filesystem (branches
not yet supported).

The acceptance test currently breaks.