#45: Basic dynamic provisioning.
The test checks that a PV gets created from a PVC, along with a
corresponding DM volume.

Actually using the volume from within a container managed by k8s is
the next step!

Also: The Delete method of the provisioner is not yet implemented, and the
build process is wasteful and woefully slow.