Fixed MacOS build.
Seems the problem was that, as the smoke tests don't run in a DinD container,
the files from the previous run linger - and RE-writing a mode 0400 password
file gives permission denied (the reason it only failed on MacOS being that
only MacOS was just running the basic smoke tests rather than the DinD-based
suite).

So this fix isn't just for MacOS - it's for any time people re-init in a
homedir that had a cluster previously.